feat(admin):  can upload images to variant

+++ apps/vendyx-api/src/business/variant/variant.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 
 import { CreateVariantInput, UpdateVariantInput } from '@/api/shared';
 import { VariantRepository } from '@/persistance/repositories';
+import { ID } from '@/persistance/types';
 
 import { clean, convertToCent } from '../shared';
 
@@ -13,29 +14,37 @@ export class VariantService {
     return this.variantRepository.findById(id);
   }
 
-  create(productId: string, input: CreateVariantInput) {
-    const { optionValues, ...rest } = input;
+  create(productId: ID, input: CreateVariantInput) {
+    const { optionValues, assetId, ...rest } = input;
 
     return this.variantRepository.insert({
       ...clean(rest),
       salePrice: convertToCent(input.salePrice),
       comparisonPrice: input.comparisonPrice ? convertToCent(input.comparisonPrice) : undefined,
       costPerUnit: input.costPerUnit ? convertToCent(input.costPerUnit) : undefined,
+      asset: assetId ? { connect: { id: assetId } } : undefined,
       product: { connect: { id: productId } },
       variantOptionValues: { create: optionValues?.map(v => ({ optionValueId: v })) }
     });
   }
 
-  async update(id: string, input: UpdateVariantInput) {
-    const { optionValues, ...rest } = input;
+  async update(id: ID, input: UpdateVariantInput) {
+    const { optionValues, assetId, ...rest } = input;
 
-    await this.variantRepository.removeOptionValuesNotIn(id, optionValues ?? []);
+    if (optionValues) {
+      await this.variantRepository.removeOptionValuesNotIn(id, optionValues ?? []);
+    }
 
     return this.variantRepository.update(id, {
       ...clean(rest),
       salePrice: input.salePrice ? convertToCent(input.salePrice) : undefined,
       comparisonPrice: input.comparisonPrice ? convertToCent(input.comparisonPrice) : undefined,
       costPerUnit: input.costPerUnit ? convertToCent(input.costPerUnit) : undefined,
+      asset: assetId
+        ? { connect: { id: assetId } }
+        : assetId === null
+        ? { disconnect: true }
+        : undefined,
       variantOptionValues: optionValues?.length
         ? {
             connectOrCreate: optionValues?.map(v => ({
